// src/enums/TooltipLocation.ts
export enum TooltipLocation {
    Auto = 0,

    BeginsTopLeft = 1,
    BeginsTopCenter = 2,
    BeginsTopRight = 3,
    BeginsCenterLeft = 4,
    BeginsCenterRight = 5,
    BeginsBottomLeft = 6,
    BeginsBottomCenter = 7,
    BeginsBottomRight = 8,

    EndsTopLeft = 9,
    EndsTopCenter = 10,
    EndsTopRight = 11,
    EndsCenterLeft = 12,
    EndsCenterRight = 13,
    EndsBottomLeft = 14,
    EndsBottomCenter = 15,
    EndsBottomRight = 16,
}